Move hooks into CoreChart
