+++ fiberplane-charts/src/hooks/index.ts
@@ -5,6 +5,3 @@ export * from "./useHandler";
 export * from "./useInteractiveControls";
 export * from "./useIntersectionObserver";
 export * from "./useMeasure";
-export * from "./useMouseControls";
-export * from "./useScales";
-export * from "./useTooltip";
